fix: various import issues
