+++ packages/backend/convex/_generated/api.d.ts
@@ -15,9 +15,9 @@ import type * as data_ml_friendRecs from "../data_ml/friendRecs.js";
 import type * as data_ml_friends from "../data_ml/friends.js";
 import type * as data_ml_universal from "../data_ml/universal.js";
 import type * as data_ml_users from "../data_ml/users.js";
+import type * as eventSeedsStatic from "../eventSeedsStatic.js";
 import type * as eventsIngest from "../eventsIngest.js";
 import type * as seed from "../seed.js";
-import type * as userAuth from "../auth.js";
 import type * as users from "../users.js";
 
 import type {
@@ -34,9 +34,9 @@ declare const fullApi: ApiFromModules<{
   "data_ml/friends": typeof data_ml_friends;
   "data_ml/universal": typeof data_ml_universal;
   "data_ml/users": typeof data_ml_users;
+  eventSeedsStatic: typeof eventSeedsStatic;
   eventsIngest: typeof eventsIngest;
   seed: typeof seed;
-  userAuth: typeof userAuth;
   users: typeof users;
 }>;
