+++ packages/backend/convex/seed.ts
@@ -1,105 +1,9 @@
 import { mutation } from './_generated/server';
-import { latLngToH3Index } from './data_ml/events';
+import { eventSeedAttendees, eventSeeds } from './eventSeedsStatic';
+
+export { eventSeedAttendees, eventSeeds };
 
 //TODO get from backend instead
-export const eventSeeds = [
-  {
-    name: 'Coffee + Homework',
-    organization: 'Pop Cafe',
-    description: 'Chill study session.',
-    location: {
-      latitude: 36.12730186736902,
-      longitude: -115.19299595922035,
-      h3Index: latLngToH3Index(36.12730186736902, -115.19299595922035),
-    },
-  },
-  {
-    name: 'ASAP Rocky Concert',
-    organization: 'ASAP Rocky',
-    description: 'Dont be dumb, pull up.',
-    location: {
-      latitude: 36.10470771447518,
-      longitude: -115.16859227452814,
-      h3Index: latLngToH3Index(36.10470771447518, -115.16859227452814),
-    },
-  },
-  {
-    name: 'Psi Rho house party',
-    organization: 'UNLV - Alpha Psi Rho',
-    description: 'no hazing, just good vibes. $10 entry for dudes.',
-    location: {
-      latitude: 36.10790291877858,
-      longitude: -115.14269190902489,
-      h3Index: latLngToH3Index(36.10790291877858, -115.14269190902489),
-    },
-  },
-  {
-    name: 'Las Vegas - First Friday',
-    organization: 'Downtown Las Vegas',
-    description:
-      'Free-admission monthly event featuring live music, art exhibits, food trucks, and vendors.',
-    location: {
-      latitude: 36.15958759998514,
-      longitude: -115.15239854806732,
-      h3Index: latLngToH3Index(36.15958759998514, -115.15239854806732),
-    },
-  },
-  {
-    name: 'St. Jimmy Panel & Conference',
-    organization: 'st. jimmy',
-    description:
-      'okay jeez, i been going thru a rough patch. going left, right thru the catacombs.',
-    location: {
-      latitude: 36.09086216508982,
-      longitude: -115.18328464909499,
-      h3Index: latLngToH3Index(36.09086216508982, -115.18328464909499),
-    },
-  },
-  {
-    name: 'LVL UP EXPO 2026',
-    organization: 'LVL UP LLC',
-    description:
-      'LVL UP EXPO is a three-day immersive gaming and pop culture convention in Las Vegas celebrating fans through esports tournaments, cosplay, art, and entertainment.',
-    location: {
-      latitude: 36.12871445143533,
-      longitude: -115.15149229313623,
-      h3Index: latLngToH3Index(36.12871445143533, -115.15149229313623),
-    },
-  },
-  {
-    name: 'Baby Keem Concert',
-    organization: 'The Cosmopolitan',
-    description: 'half past twelve i was all alone.',
-    location: {
-      latitude: 36.10987909759377,
-      longitude: -115.17538973403965,
-      h3Index: latLngToH3Index(36.10987909759377, -115.17538973403965),
-    },
-  },
-  {
-    name: 'Water Lantern Festival',
-    organization: 'SWCTA Key Club',
-    description: 'Join us at Sunset Park to litter the pond.',
-    location: {
-      latitude: 36.03809903957797,
-      longitude: -115.24699453074136,
-      h3Index: latLngToH3Index(36.03809903957797, -115.24699453074136),
-    },
-  },
-  {
-    name: 'Thrift Valley pop up shop',
-    organization: 'Thrift Valley',
-    description: 'we outta stussy dont even ask',
-    location: {
-      latitude: 36.15909747099756,
-      longitude: -115.15269829421878,
-      h3Index: latLngToH3Index(36.15909747099756, -115.15269829421878),
-    },
-  },
-];
-
-// Attendee counts per event, parallel to eventSeeds — derived from userEventPairs in seed handler
-export const eventSeedAttendees = [3, 5, 4, 4, 3, 5, 2, 4, 6];
 
 export const seed = mutation({
   args: {},
